HAWZX-AI/backend/services/vision_service.py
from transformers import ViTFeatureExtractor, ViTForImageClassification, DetrForObjectDetection
from PIL import Image
import pyautogui
import numpy as np # Adicionado para manipulação de imagem se necessário
import easyocr # Para OCR

class VisionService:
    def __init__(self):
        # Carregar modelos na inicialização (placeholders)
        # O ideal é carregar apenas um modelo ViT para classificação de jogo
        # e modelos DETR específicos para cada jogo detectado dinamicamente
        print("Initializing VisionService...")
        try:
            self.game_classifier_feature_extractor = ViTFeatureExtractor.from_pretrained("google/vit-base-patch16-224")
            self.game_classifier_model = ViTForImageClassification.from_pretrained("google/vit-base-patch16-224")
            # Exemplo de um modelo DETR, você precisaria de modelos treinados para seus objetos específicos
            # self.object_detector_feature_extractor = DetrFeatureExtractor.from_pretrained("facebook/detr-resnet-50")
            # self.object_detector_model = DetrForObjectDetection.from_pretrained("facebook/detr-resnet-50")
            self.ocr_reader = easyocr.Reader(['en', 'pt']) # Idiomas que você planeja suportar

            self.supported_games = {
                "game_a": {"model_path": "path/to/game_a_detr_model"},
                "game_b": {"model_path": "path/to/game_b_detr_model"},
            }
            print("VisionService initialized successfully with placeholder models.")
        except Exception as e:
            print(f"Error initializing VisionService models: {e}. Some models might not be loaded.")
            self.game_classifier_feature_extractor = None
            self.game_classifier_model = None
            self.ocr_reader = None


    def capture_screen(self):
        """Captura a tela inteira."""
        print("Capturing screen...")
        screenshot = pyautogui.screenshot()
        return screenshot

    def detect_game(self, screenshot: Image.Image):
        """Detecta o jogo ativo a partir da captura de tela."""
        if not self.game_classifier_model or not self.game_classifier_feature_extractor:
            print("Game classifier not loaded. Skipping game detection.")
            return "unknown"
        
        print("Detecting game...")
        # Preprocessar a imagem para o modelo ViT
        inputs = self.game_classifier_feature_extractor(images=screenshot, return_tensors="pt")
        outputs = self.game_classifier_model(**inputs)
        logits = outputs.logits
        # A classe detectada seria a de maior probabilidade
        predicted_class_idx = logits.argmax(-1).item()
        # Mapeie o índice da classe para o nome do jogo. Isso requer um mapeamento pré-definido.
        # Por exemplo: self.game_classifier_model.config.id2label[predicted_class_idx]
        
        # Placeholder para o nome do jogo detectado
        detected_game = "game_a" # Simular detecção
        print(f"Game detected: {detected_game}")
        return detected_game

    def detect_objects(self, screenshot: Image.Image, game_name: str):
        """Detecta objetos específicos do jogo na captura de tela."""
        # Carregar o modelo DETR específico para 'game_name'
        # if game_name not in self.supported_games or not self.object_detector_model:
        #     print(f"No object detection model for {game_name} or model not loaded.")
        #     return []

        print(f"Detecting objects for {game_name}...")
        # Placeholder para a lógica de detecção de objetos
        # inputs = self.object_detector_feature_extractor(images=screenshot, return_tensors="pt")
        # outputs = self.object_detector_model(**inputs)
        # Processar as saídas para obter bounding boxes e labels

        # Exemplo de retorno de objetos detectados (placeholder)
        detected_objects = [
            {"label": "health_bar", "box": [10, 10, 100, 20]},
            {"label": "enemy", "box": [200, 150, 250, 200]}
        ]
        print(f"Detected {len(detected_objects)} objects.")
        return detected_objects

    def extract_text(self, screenshot: Image.Image, rois: list):
        """Extrai texto de Regiões de Interesse (ROIs) usando OCR."""
        if not self.ocr_reader:
            print("OCR reader not loaded. Skipping text extraction.")
            return {}

        print("Extracting text from ROIs...")
        text_data = {}
        for roi in rois:
            # roi_image = screenshot.crop(roi["box"])
            # results = self.ocr_reader.readtext(np.array(roi_image))
            # text_data[roi["label"]] = " ".join([res[1] for res in results])
            text_data[roi["label"]] = f"text_from_{roi['label']}" # Placeholder
        print(f"Extracted text: {text_data}")
        return text_data

    def analyze_screen(self):
        """Orquestra a análise completa da tela."""
        screenshot = self.capture_screen()
        game = self.detect_game(screenshot)
        
        # Em um cenário real, `detect_objects` e `extract_text` dependeriam do 'game' detectado
        # e de configurações específicas para cada jogo.
        
        # Placeholder para ROIs baseadas no jogo
        # rois_for_game = self.get_rois_for_game(game) 
        
        objects = self.detect_objects(screenshot, game)
        
        # Usar ROIs dos objetos detectados para extrair texto
        text_rois = [obj for obj in objects if "text_region" in obj.get("label", "")] # Exemplo
        text_data = self.extract_text(screenshot, text_rois)

        return {"game": game, "objects": objects, "text": text_data}

