HAWZX-AI/backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
sentencepiece
sacremoses
faiss-cpu
sentence-transformers
pyttsx3
SpeechRecognition
pyautogui
Pillow
easyocr
requests